move protocol updates to sub view model
